requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
Django==4.2.4
django-ckeditor==6.7.0
django-js-asset==2.1.0
django-sendgrid-v5==1.2.2
djangorestframework==3.14.0
idna==3.4
optimized-image==0.0.1
Pillow==10.0.0
psycopg2-binary==2.9.7
pycparser==2.21
python-http-client==3.3.7
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
tinify==1.6.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
